Added crudl integration tests
