+++ integration/client_crudl_test.go
@@ -1,11 +1,8 @@
-package main
+package integration
 
 import (
 	"encoding/json"
-	"flag"
 	"fmt"
-	"os/exec"
-	"regexp"
 	"strings"
 	"testing"
 )
@@ -20,36 +17,6 @@ type MobileClientJson struct {
 	Spec MobileClientSpec
 }
 
-var namespace = flag.String("namespace", "", "Openshift namespace (most often Project) to run our integration tests in")
-var name = flag.String("name", "", "Client name to be created")
-var executable = flag.String("executable", "", "Executable under test")
-
-type ValidationFunction = func(output []byte, err error) (bool, []string)
-
-func EmptyValidation(output []byte, err error) (bool, []string) {
-	return true, []string{}
-}
-
-func VNoErr(output []byte, err error) (bool, []string) {
-	return err == nil, []string{fmt.Sprintf("%s", err)}
-}
-
-func VIsErr(output []byte, err error) (bool, []string) {
-	return err != nil, []string{fmt.Sprintf("%s", err)}
-}
-
-func VRegex(pattern string) func(output []byte, err error) (bool, []string) {
-	return func(output []byte, err error) (bool, []string) {
-		matched, errMatch := regexp.MatchString(pattern, fmt.Sprintf("%s", output))
-		if errMatch != nil {
-			return false, []string{fmt.Sprintf("Error in regexp %s when trying to match %s", errMatch, pattern)}
-		}
-		if !matched {
-			return false, []string{fmt.Sprintf("Expected combined output matching %s", pattern)}
-		}
-		return true, []string{}
-	}
-}
 func VMobileClientJson(name string, clientType string) func(output []byte, err error) (bool, []string) {
 	return func(output []byte, err error) (bool, []string) {
 		var parsed MobileClientJson
@@ -67,49 +34,7 @@ func VMobileClientJson(name string, clientType string) func(output []byte, err e
 	}
 }
 
-func All(vs ...ValidationFunction) ValidationFunction {
-	return func(output []byte, err error) (bool, []string) {
-		for _, v := range vs {
-			r, o := v(output, err)
-			if !r {
-				return r, o
-			}
-		}
-		return true, []string{}
-	}
-}
-
-type CmdDesc struct {
-	executable string
-	Arg        []string
-	Validator  ValidationFunction
-}
-
-func (c CmdDesc) Add(arg ...string) CmdDesc {
-	return CmdDesc{c.executable, append(c.Arg, arg...), c.Validator}
-}
-
-func (c CmdDesc) Complying(validator ValidationFunction) CmdDesc {
-	return CmdDesc{c.executable, c.Arg, All(c.Validator, validator)}
-}
-
-func (c CmdDesc) Run(t *testing.T) ([]byte, error) {
-	t.Log(c.Arg)
-	cmd := exec.Command(c.executable, c.Arg...)
-	output, err := cmd.CombinedOutput()
-	t.Log(fmt.Sprintf("%s\n", output))
-	v, errs := c.Validator(output, err)
-	if !v {
-		t.Fatal(errs)
-	}
-	return output, err
-}
-
-func ValidatedCmd(executable string, arg ...string) CmdDesc {
-	return CmdDesc{executable, arg, EmptyValidation}
-}
-
-func TestPositive(t *testing.T) {
+func TestClientJson(t *testing.T) {
 
 	clientTypes := []string{
 		"cordova",
@@ -117,18 +42,24 @@ func TestPositive(t *testing.T) {
 		"android",
 	}
 
+	name := fmt.Sprintf("%s-mobile-crud-test-entity", *prefix)
 	m := ValidatedCmd(*executable, fmt.Sprintf("--namespace=%s", *namespace), "-o=json")
+	oc := ValidatedCmd("oc", fmt.Sprintf("--namespace=%s", *namespace), "-o=json")
 	for _, clientType := range clientTypes {
 		t.Run(clientType, func(t *testing.T) {
-			expectedId := strings.ToLower(fmt.Sprintf("%s-%s", *name, clientType))
-			notExists := All(VIsErr, VRegex(".*Error: failed to get.*"))
-			exists := All(VNoErr, VMobileClientJson(*name, clientType))
+			expectedId := strings.ToLower(fmt.Sprintf("%s-%s", name, clientType))
+
+			notExists := All(VIsErr, VRegex(fmt.Sprintf(".*\"%s\" not found.*", expectedId)))
+			exists := All(VNoErr, VMobileClientJson(name, clientType))
+
 			m.Add("get", "client", expectedId).Complying(notExists).Run(t)
-			m.Add("create", "client", *name, clientType).Complying(exists).Run(t)
+			oc.Add("get", "mobileclient", expectedId).Complying(notExists).Run(t)
+			m.Add("create", "client", name, clientType).Complying(exists).Run(t)
 			m.Add("get", "client", expectedId).Complying(exists).Run(t)
+			oc.Add("get", "mobileclient", expectedId).Complying(exists).Run(t)
 			m.Add("delete", "client", expectedId).Complying(VNoErr).Run(t)
 			m.Add("get", "client", expectedId).Complying(notExists).Run(t)
-
+			oc.Add("get", "mobileclient", expectedId).Complying(notExists).Run(t)
 		})
 	}
 }
